Preparing to push this to pypi. Almost done.
Logically, everything is the same, but added a main function in game.py
for pypi.

Even now game.py is all anyone needs to run this (in python3).

My computer is currently on python 2.7.6, and twine throws an error
about urllib3 not supporting SSL. So, I need to upgrade my py
installation to be able to upload.